Tentativo di Refactoring per ottenere un livello di coupling minore tramite Interfaccia Board.java
