+++ src/main/java/it/units/gomokusdm/CLIController.java
@@ -28,7 +28,7 @@ public class CLIController {
         this.reader = new BufferedReader(new InputStreamReader(inputStream));
         this.player1 = new Player("", Stone.BLACK);
         this.player2 = new Player("", Stone.WHITE);
-        this.board = new Board();
+        this.board = new BoardImplementation();
         try {
             this.game = new Game(board, player1, player2);
         } catch (Exception e) {
