+++ src/test/java/it/units/gomokusdm/CLIControllerTest.java
@@ -10,7 +10,7 @@ public class CLIControllerTest {
 
     @Test
     void test() {
-        Board board = new Board();
+        Board board = new BoardImplementation();
         board.setCell(Stone.BLACK, new Coordinates(0, 0));
         CLIController cli = CLIController.createInstance(System.out, System.in);
         cli.printBoard();
